+++ packages/core-js/actual/element/remove.js
@@ -1,0 +1,4 @@
+'use strict';
+var parent = require('../../stable/element/remove');
+
+module.exports = parent;