+++ packages/core-js/modules/web.element.remove.js
@@ -1,0 +1,34 @@
+'use strict';
+var $ = require('../internals/export');
+var uncurryThis = require('../internals/function-uncurry-this');
+var globalThis = require('../internals/global-this');
+var anElement = require('../internals/an-element');
+var wellKnownSymbol = require('../internals/well-known-symbol');
+var create = require('../internals/object-create');
+var defineProperty = require('../internals/object-define-property').f;
+
+if (globalThis.Element) {
+  var BASIC = !!Element.prototype.remove;
+  var removeChild;
+  if (Element.prototype.removeChild) removeChild = uncurryThis(Element.prototype.removeChild);
+
+  // `Element.remove` method
+  // https://dom.spec.whatwg.org/#dom-childnode-remove
+  $({ target: 'Element', proto: true, forced: !BASIC }, {
+    remove: function remove() {
+      anElement(this);
+      if (this.parentNode === null) return;
+      if (removeChild) removeChild(this.parentNode, this);
+      // Old Android Browser do not have removeChild.
+      else this.outerHTML = '';
+    }
+  });
+
+  if (Element.prototype[wellKnownSymbol('unscopables')] === undefined) {
+    defineProperty(Element.prototype, wellKnownSymbol('unscopables'), {
+      configurable: true,
+      value: create(null)
+    });
+  }
+  Element.prototype[wellKnownSymbol('unscopables')].remove = true;
+}